requirements.txt
loguru==0.7.2
requests~=2.32.3
nicegui==2.13.0
ruff==0.3.5
langgraph==0.3.11
pydantic~=2.8.2
python-dotenv==1.0.1
langchain==0.3.6
langchain-core~=0.3.12
langchain_community==0.3.4
grpcio==1.60.1
google-genai==1.2.0
google-cloud-storage==3.0.0
setuptools~=75.8.0
protobuf~=4.25.6
tqdm~=4.67.1
tenacity~=9.0.0
pymongo==4.11.2
fastapi==0.115.8
uvicorn==0.34.0
httpx~=0.28.1
langchain-mongodb==0.5.0
langchain-google-genai~=2.0.9
markdown-it-py[plugins]~=3.0.0
beautifulsoup4==4.13.3
slowapi~=0.1.9
redis==5.2.1